Extract enums used by nodes into separate module
