+++ src/pygerber/gerberx3/ast/nodes/attribute/TA.py
@@ -2,12 +2,12 @@
 
 from __future__ import annotations
 
-from enum import Enum
 from typing import TYPE_CHECKING, Callable, List, Literal, Optional
 
 from pydantic import Field
 
 from pygerber.gerberx3.ast.nodes.base import Node
+from pygerber.gerberx3.ast.nodes.enums import AperFunction
 
 if TYPE_CHECKING:
     from typing_extensions import Self
@@ -34,48 +34,6 @@ class TA_UserName(TA):  # noqa: N801
     ) -> Callable[[Self], None]:
         """Get callback function for the node."""
         return visitor.on_ta_user_name
-
-
-class AperFunction(Enum):
-    """Enum representing possible AperFunction values."""
-
-    ViaDrill = "ViaDrill"
-    BackDrill = "BackDrill"
-    ComponentDrill = "ComponentDrill"
-    MechanicalDrill = "MechanicalDrill"
-    CastellatedDrill = "CastellatedDrill"
-    OtherDrill = "OtherDrill"
-    ComponentPad = "ComponentPad"
-    SMDPad = "SMDPad"
-    BGAPad = "BGAPad"
-    ConnectorPad = "ConnectorPad"
-    HeatsinkPad = "HeatsinkPad"
-    ViaPad = "ViaPad"
-    TestPad = "TestPad"
-    CastellatedPad = "CastellatedPad"
-    FiducialPad = "FiducialPad"
-    ThermalReliefPad = "ThermalReliefPad"
-    WasherPad = "WasherPad"
-    AntiPad = "AntiPad"
-    OtherPad = "OtherPad"
-    Conductor = "Conductor"
-    EtchedComponent = "EtchedComponent"
-    NonConductor = "NonConductor"
-    CopperBalancing = "CopperBalancing"
-    Border = "Border"
-    OtherCopper = "OtherCopper"
-    ComponentMain = "ComponentMain"
-    ComponentOutline = "ComponentOutline"
-    ComponentPin = "ComponentPin"
-    Profile = "Profile"
-    Material = "Material"
-    NonMaterial = "NonMaterial"
-    Other = "Other"
-
-    def __repr__(self) -> str:
-        return f"{self.__class__.__name__}.{self.name}"
-
-    __str__ = __repr__
 
 
 class TA_AperFunction(TA):  # noqa: N801
