+++ src/pygerber/gerberx3/ast/nodes/properties/AS.py
@@ -2,22 +2,15 @@
 
 from __future__ import annotations
 
-from enum import Enum
 from typing import TYPE_CHECKING, Callable
 
 from pygerber.gerberx3.ast.nodes.base import Node
+from pygerber.gerberx3.ast.nodes.enums import AxisCorrespondence
 
 if TYPE_CHECKING:
     from typing_extensions import Self
 
     from pygerber.gerberx3.ast.visitor import AstVisitor
-
-
-class AxisCorrespondence(Enum):
-    """Represents axis correspondence."""
-
-    AX_BY = "AXBY"
-    AY_BX = "AYBX"
 
 
 class AS(Node):
