+++ src/pygerber/gerberx3/ast/nodes/properties/FS.py
@@ -2,9 +2,10 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Callable, Literal
+from typing import TYPE_CHECKING, Callable
 
 from pygerber.gerberx3.ast.nodes.base import Node
+from pygerber.gerberx3.ast.nodes.enums import CoordinateMode, Zeros
 
 if TYPE_CHECKING:
     from typing_extensions import Self
@@ -15,8 +16,8 @@ if TYPE_CHECKING:
 class FS(Node):
     """Represents FS Gerber extended command."""
 
-    zeros: Literal["L", "T"]
-    coordinate_mode: Literal["A", "I"]
+    zeros: Zeros
+    coordinate_mode: CoordinateMode
 
     x_integral: int
     x_decimal: int
