+++ src/pygerber/gerberx3/ast/nodes/properties/IP.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Callable
 
 from pygerber.gerberx3.ast.nodes.base import Node
+from pygerber.gerberx3.ast.nodes.enums import ImagePolarity
 
 if TYPE_CHECKING:
     from typing_extensions import Self
@@ -15,7 +16,7 @@ if TYPE_CHECKING:
 class IP(Node):
     """Represents IP Gerber extended command."""
 
-    polarity: str
+    polarity: ImagePolarity
 
     def visit(self, visitor: AstVisitor) -> None:
         """Handle visitor call."""
